+++ packages/templating/vitest.config.js
@@ -8,6 +8,7 @@ export default defineConfig({
       enabled: true,
       headless: true,
       provider: 'playwright',
+      screenshotFailures: false,
       instances: [
         {
           browser: 'chromium',